requirements.txt
apache-airflow-providers-postgres==5.6.1
pandas==1.5.3
psycopg2-binary==2.9.9
Faker==19.12.0
numpy==1.23.5
xmltodict==0.13.0